3D空間でのレイトレに向けて色々実装しました。 スラリンラボのカメラ付きレイトレの章を一通りビルドが通る様に実装したレベル。 実行はできるが真っ暗なので、GPUデバッガでデバッグする必要あり。 保存のためにコミット。

+++ Struct.h
@@ -335,7 +335,7 @@ struct ConstBufferDataMultiPath {
 //ModelDataManagerとObject3D間のデータの受け渡し用関数
 struct Object3DDeliveryData {
 	vector<Vertex> vertex;
-	vector<unsigned short> index;
+	vector<UINT> index;
 	ConstBufferDataB1 constBufferDataB1;
 };
